fix(backend): do not drop a reference on the EVP_PKEY in sig path
There is no bumpref on the EVP_PKEY, so the RSA key will get the
refcount and decref and the end of the function - the EVP_PKEY_free
is incorrect since it just frees the passed in reference on the
gen_sig_data. This code path appears to be have been corrected
in the upstream/latest.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>